segregate EthApiClient into smaller interfaces.

+++ jvm-libs/linea/clients/interfaces/src/main/kotlin/linea/ethapi/LineaApiClient.kt
@@ -2,8 +2,7 @@ package linea.ethapi
 
 import linea.domain.TransactionForEthCall
 import tech.pegasys.teku.infrastructure.async.SafeFuture
-import java.math.BigInteger
 
 interface LineaApiClient {
-  fun lineaEstimateGas(transaction: TransactionForEthCall): SafeFuture<BigInteger>
+  fun lineaEstimateGas(transaction: TransactionForEthCall): SafeFuture<ULong>
 }
